requirements.txt
uvicorn
pydantic
fastapi
psycopg2-binary
sqlalchemy>=2.0.0
pytest
httpx
pytest-asyncio
redis-server
